// webapp/src/core/types.ts
export type ID = string;

/**
 * Represents time in Milliseconds.
 * All time values in the core engine use this unit.
 */
export type TimeMs = number;

export interface Point {
    x: number;
    y: number;
}

export interface Size {
    width: number;
    height: number;
}

// The rect point represents the top-left corner.
export interface Rect extends Point, Size { }
// ...
export interface DeviceFrame {
    id: ID;
    name: string;
    imageUrl: string;
    thumbnailUrl: string;
    // The inner screen rectangle in the frame image (relative to image 0,0)
    // Used to calculate border thickness ratios
    screenRect: Rect;
    // Total size of the frame image
    size: Size;
    borderData: FrameBorderData;
    customScaling?: FrameScalingConfig;
}

export interface FrameScalingConfig {
    vertical: SliceSegment[];
    horizontal: SliceSegment[];
}

export interface SliceSegment {
    start: number;
    end: number;
    scalable: boolean;
}

export interface FrameBorderData {
    // Ratios of border thickness to total size (0..1)
    top: number;
    bottom: number;
    left: number;
    right: number;
}